requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
Django==2.2
django-cors-headers==2.1.0
django-notifs==3.0.0
django-templated-mail==1.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
gevent==21.1.2
greenlet==1.0.0
kombu==4.6.11
pika==0.11.2
PyJWT==1.6.1
pytz==2017.3
simplejson==3.13.2
six==1.11.0
uWSGI==2.0.19.1
vine==1.3.0
